refactor: move to `Tags` category
