+++ src/commands/slash/tags/Create.ts
@@ -1,10 +1,10 @@
-import type BaseClient from '../../../../lib/BaseClient.js';
-import Command from '../../../../lib/structures/Interaction.js';
+import type BaseClient from '../../../lib/BaseClient.js';
+import Command from '../../../lib/structures/Interaction.js';
 import { ActionRowBuilder, ModalBuilder, TextInputBuilder } from '@discordjs/builders';
 import { InteractionType, TextInputStyle } from 'discord-api-types/v10';
 import { ChatInputCommandInteraction, InteractionCollector, ModalSubmitInteraction } from 'discord.js';
 import { inlineCode } from '@discordjs/formatters';
-import { slugify } from '../../../../lib/utils/Function.js';
+import { slugify } from '../../../lib/utils/Function.js';
 import { nanoid } from 'nanoid';
 
 export default class extends Command {
@@ -12,7 +12,7 @@ export default class extends Command {
 		super(client, {
 			name: 'tags create',
 			description: 'Create a new server tag.',
-			category: 'Manage',
+			category: 'Tags',
 			memberPermissions: ['ManageGuild'],
 			guildOnly: true
 		});
