+++ src/commands/slash/tags/Delete.ts
@@ -1,10 +1,10 @@
-import type BaseClient from '../../../../lib/BaseClient.js';
-import Command from '../../../../lib/structures/Interaction.js';
+import type BaseClient from '../../../lib/BaseClient.js';
+import Command from '../../../lib/structures/Interaction.js';
 import { ActionRowBuilder, ButtonBuilder } from '@discordjs/builders';
 import { ButtonStyle, ComponentType } from 'discord-api-types/v10';
 import type { AutocompleteInteraction, ButtonInteraction, ChatInputCommandInteraction } from 'discord.js';
 import { inlineCode } from '@discordjs/formatters';
-import { shuffleArray } from '../../../../lib/utils/Function.js';
+import { shuffleArray } from '../../../lib/utils/Function.js';
 import { nanoid } from 'nanoid';
 
 export default class extends Command {
@@ -12,7 +12,7 @@ export default class extends Command {
 		super(client, {
 			name: 'tags delete',
 			description: 'Delete a server tag.',
-			category: 'Manage',
+			category: 'Tags',
 			memberPermissions: ['ManageGuild'],
 			guildOnly: true
 		});
