+++ src/commands/slash/tags/Pin.ts
@@ -1,13 +1,16 @@
 import type BaseClient from '../../../lib/BaseClient.js';
 import Command from '../../../lib/structures/Interaction.js';
 import type { AutocompleteInteraction, ChatInputCommandInteraction } from 'discord.js';
+import { inlineCode } from '@discordjs/formatters';
+import { shuffleArray } from '../../../lib/utils/Function.js';
 
 export default class extends Command {
 	public constructor(client: BaseClient) {
 		super(client, {
-			name: 'tag',
-			description: 'Send a server tag.',
-			category: 'Utility',
+			name: 'tags pin',
+			description: 'Pin a server tag.',
+			category: 'Tags',
+			memberPermissions: ['ManageGuild'],
 			guildOnly: true
 		});
 	}
@@ -23,7 +26,15 @@ export default class extends Command {
 		const filtered = prisma?.tags.find(({ slug }) => slug === name);
 		if (!filtered) return interaction.reply({ content: 'The tag name doesn\'t exist.', ephemeral: true });
 
-		return interaction.reply({ content: filtered.content });
+		const pinned = prisma?.tags.filter(({ hoisted }) => hoisted);
+		if (pinned!.length >= 25) return interaction.reply({ content: 'Unable to pin more than 25 tags.', ephemeral: true });
+
+		await this.client.prisma.tag.update({
+			where: { id: filtered.id },
+			data: { hoisted: true }
+		});
+
+		return interaction.reply({ content: `Tag ${inlineCode(name)} has been pinned.`, ephemeral: true });
 	}
 
 	public override async autocomplete(interaction: AutocompleteInteraction<'cached'>) {
@@ -37,14 +48,8 @@ export default class extends Command {
 		const choices = prisma?.tags.filter(({ name }) => name.toLowerCase().includes(focused.toLowerCase()));
 		if (!choices?.length) return interaction.respond([]);
 
-		let respond = choices.filter(({ hoisted }) => hoisted).map(({ name, slug }) => ({ name, value: slug }));
+		const respond = choices.map(({ name, slug }) => ({ name, value: slug }));
 
-		if (focused.length) {
-			respond = choices.map(({ name, slug }) => ({ name, value: slug }));
-
-			return interaction.respond(respond.slice(0, 25));
-		}
-
-		return interaction.respond(respond.slice(0, 25));
+		return interaction.respond(shuffleArray(respond.slice(0, 25)));
 	}
 }
