+++ src/commands/slash/tags/Rename.ts
@@ -1,18 +1,18 @@
-import type BaseClient from '../../../../lib/BaseClient.js';
-import Command from '../../../../lib/structures/Interaction.js';
+import type BaseClient from '../../../lib/BaseClient.js';
+import Command from '../../../lib/structures/Interaction.js';
 import { ActionRowBuilder, ModalBuilder, TextInputBuilder } from '@discordjs/builders';
 import { InteractionType, TextInputStyle } from 'discord-api-types/v10';
 import { AutocompleteInteraction, ChatInputCommandInteraction, InteractionCollector, ModalSubmitInteraction } from 'discord.js';
 import { inlineCode } from '@discordjs/formatters';
-import { shuffleArray } from '../../../../lib/utils/Function.js';
+import { shuffleArray, slugify } from '../../../lib/utils/Function.js';
 import { nanoid } from 'nanoid';
 
 export default class extends Command {
 	public constructor(client: BaseClient) {
 		super(client, {
-			name: 'tags edit',
-			description: 'Edit a server tag.',
-			category: 'Manage',
+			name: 'tags rename',
+			description: 'Rename a server tag.',
+			category: 'Tags',
 			memberPermissions: ['ManageGuild'],
 			guildOnly: true
 		});
@@ -32,15 +32,15 @@ export default class extends Command {
 		const modalId = nanoid();
 		const modal = new ModalBuilder()
 			.setCustomId(modalId)
-			.setTitle('Edit a server tag')
+			.setTitle('Rename a server tag')
 			.setComponents(new ActionRowBuilder<TextInputBuilder>()
 				.setComponents(new TextInputBuilder()
-					.setCustomId('content')
-					.setStyle(TextInputStyle.Paragraph)
-					.setLabel('Content')
-					.setValue(filtered.content)
+					.setCustomId('title')
+					.setStyle(TextInputStyle.Short)
+					.setLabel('Title')
+					.setValue(filtered.name)
 					.setRequired(true)
-					.setMaxLength(2000)));
+					.setMaxLength(100)));
 
 		await interaction.showModal(modal);
 
@@ -48,14 +48,20 @@ export default class extends Command {
 		const collector = new InteractionCollector(this.client, { filter, interactionType: InteractionType.ModalSubmit, max: 1 });
 
 		collector.on('collect', async (i) => {
-			const content = i.fields.getTextInputValue('content');
+			const title = i.fields.getTextInputValue('title');
+
+			const exist = prisma?.tags.some(({ slug }) => slug === slugify(title));
+			if (exist) return void i.reply({ content: 'A tag with that name already exists.', ephemeral: true });
 
 			await this.client.prisma.tag.update({
 				where: { id: filtered.id },
-				data: { content }
+				data: {
+					slug: slugify(title),
+					name: title
+				 }
 			});
 
-			return void i.reply({ content: `Tag ${inlineCode(name)} has been edited.`, ephemeral: true });
+			return void i.reply({ content: `Tag ${inlineCode(name)} has been renamed to ${inlineCode(slugify(title))}.`, ephemeral: true });
 		});
 	}
